Project 1 Final
This Project discusses how federal, non-military employment was im-
pacted by the transition from President George Bush to President Barack
Obama to learn about the impact of the dierence in policy between the
two administrations on the federal government. In the First section the
Project will look at the campaign promises from each and then look at the
actual policies applied by each and the Key events to see if that aected
the Federal employment strategy.After that the project explains the data
set provided, what it means, how was it cleaned and what are the assump-
tions and Imputations. In the Third section some agencies are selected for
analysis and visualization using appropiate methods. The Fourth and the
last section has the conclusion of the project, and discussed the limitation
to work, and how can this report taken for further research